Move tests for as_type() function
